A better way to do test data scripts. Move things into files that get imported, so the files doing the actual work need not worry about interactive mode shenanigans.
